Refactor/renaming
