+++ aggregate_store.go
@@ -8,7 +8,7 @@ import (
 )
 
 // An AggregateStore is a read/write store for aggregates.
-// See package `aggregatestore` for various implementations.
+// See package `aggregatestore` for implementations.
 type AggregateStore[E Entity] interface {
 	NewAggregate() (*Aggregate[E], error)
 	Load(ctx context.Context, id typeid.TypeID, opts LoadAggregateOptions) (*Aggregate[E], error)
